requirements.txt
openai==0.28.0
langchain==0.0.350
chromadb==0.4.22
streamlit==1.31.0
pypdf==3.17.4
python-dotenv==1.0.0
tiktoken==0.5.2